init: Weaken property override security for the init extension
Sometimes we need to override ro.* properties by using our vendor init
extension.

Previously there was a security check which was blocking that.
To resolve the issue, we need to weaken the security check during the
execution of our vendor init extension.

This is safe because the vendor init extension gets executed as part of init
construction and it is considered a trusted system component.

Change-Id: I6198b453745cb92c65d3e3d49e3262354cddd2a2
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>

Former-commit-id: 15aafc0c179544dd173a7afdce8d4d24b19219e7

Former-commit-id: d5ce44bfdf41cbe51f7432a04409bb11147b35a4

Former-commit-id: ff0763ca639d4987d1130f9838474b935f9b9950

Former-commit-id: 760b30eff74a6e212f57b09be8505ae8283ef3c8

Former-commit-id: 259b97b043c0769d4c8886c9ace9686ca2188536

Former-commit-id: b2022a67efa74fb7bc00e52fdac4a545599104a2
Former-commit-id: 16a827cae1ccc17f2b3fc0b35b979619c1c195be
Former-commit-id: 3f64b3f7a1c4ba677bc02d172f8565e00ad803d7